Fix being unable to find native module in built app.
This aligns the relative path to the native module from
active-game-manager between the dev and built versions. Kinda ugly but
it's our only one and this is the only place it's used so? Find I guess.

Had to adjust how the build works a bit since this really needs to be a
dynamic import to make sure things align. We no longer need to specify
the chunk stuff in the webpack config, and can instead do it at the
place where it is imported (not sure I really prefer it, but webpack
really wants to give it a dynamic chunk name otherwise).