Ignore failures for game reconciliation from request interleaving.
Funnily enough this happened to a single person twice in a row,
they must be profoundly unlucky or something :) In any case, this
shouldn't be harmful: the request that gets there first will calculate
the results the same as this one had. I verified that when this occurred
on the server, calculations still made it through.