Make matchmaking service more resilient to disconnects and maliciousness:
- Don't let users cancel game loading of games they aren't in
- Pass through a cancel token to notify GameLoader when people leave
- If the server tells us the game doesn't exist, quit it locally

The Matchmaking service is kind of hard to follow as it is currently,
since control flow runs back and forth through a bunch of different
places/classes. I think the whole GameLoader setup (as well as
MatchAcceptor and other stuff) needs kind of a rethink. The matchmaking
service would be better off as a standard HTTP API with a backing
service (like Parties) anyway. Definitely think we can simplify this
stuff if we actually embrace the server-controlled game instances thing
instead of just half-embracing it.