Rework the email verification notification.
The previous version relied on some weird behavior of ScrollableContent
to not cause a scrollbar on the entire app (since the content is sized
by `calc(100% - appBarHeight)`, and this notification was adding 32px of
extra height. I have no idea how that worked before but OKAY :) ).

This feels like a lot of text to fit in this area so I cut a bit of it,
but it mostly looks okay. Probably we should think more on this design
at some point.