Fix blank chat user list after changing views.
Needed a few things:

1) shouldComponentUpdate removed, was challenging to find the right set
of conditions to make this work properly, and this render, even with 10s
of thousands of users, is like ~1ms now in dev mode. Should be fine.

2) Resize listener added to properly resize the virtualized list so its
scroll area is correct (and it renders the right amount of things

3) Needed to tell react-virtualized to recompute heights when the users
change (this was resulting in e.g. our offline overline being sized to
8px tall and invisible on the first update sometimes

Fixes #642.