Remove admin invites UI.
This was broken in the wouter transition and I didn't really consider it
useful to fix (everyone that signed up for an invite has already been
invited on the public server). Gonna leave the APIs and DB stuff for
now but we can clean that up eventually.

This also fixes hot reloading in the admin panels (using the same
technique as the Dev stuff).