Add special logic for matching "high ranked" players.
This makes it so players that are above a certain rating threshold (top
1% of ranked players as of this commit) will be allowed to expand their
search bounds infinitely, and are not required to have a symmetric match
with potential opponents (that is, as long as their opponent is in their
search bounds, they do not have to be in the opponent's search bounds).

This should hopefully prevent some players from getting so highly ranked
that they can't find a match.

This commit also fixes the matchmaker not considering whether matches
were symmetric (it simply wasn't implemented or tested for previously,
whoops!).