Store information about matchmaking queue size in redis.
This is stored in a sorted set keyed by matchmaking type, e.g.:
matchmaking:1v1:queue

In JSON format:
{ time: <timestamp>, size: <size> }

On server start, we setup a timer to do this every 30 seconds, and clear
any entries in the set older than 3 months (may need to adjust this).

To retrieve the current values, you can do:

zrange matchmaking:1v1:queue 0 9999999999999

or if you need a specific range:

zrange matchmaking:1v1:queue <startTime> <endTime>

There's no built-in UI in the app for querying/displaying this info yet,
I'll probably add it when I add more matchmaking debug stuff in the
admin panel. (And maybe we could summarize this info for displaying to
users as well)