Use Inter/Cabin typescale.
These are both variable fonts so we get a bunch more usable weights for
a smaller download, and having a medium-weight secondary font is pertty
useful for the sizes we use it at.

Some of the places that still use the old type styles stand out a bit
more now, will adjust those the main layout ones in the next commit.