Remove the arm64 build of the server for now.
This makes it take *forever* on my machine, perhaps if we can offload
this to the cloud or I pick up a small arm machine to do this on it
would be worth it. For now I don't anticipate us needing to run this on
an arm machine (unlike the rally-point stuff).