Remove ScrollableContent from chat + whispers.
The MessageList thing I made for these is probably enough of an example
to replace any other ScrollableContent that needs info about scroll
position (although honestly very few places are as complex as the chat
stuff).

This also moved a number of things to typescript (including the chat
message records, which can probably be made a bit shorter/less
boilerplate-y in the future).

Fixes the mysterious 1px grey line at the bottom of chat windows (still
present elsewhere in the app though :( ).