Rework how chat message styling works to remove flexbox + make copy and pasted messages nice.
This also removes the shouldComponentUpdate in MessageList. I was pretty
sure it was unnecessary in the previous chat changes but I added it back
in when I was having infinitely-looping scroll updates. I ended up
fixing that but had left the sCU method in, but in working on this stuff
it got very annoying to have it preventing hot reloads all the time.

Anyway, I don't think it did much before since we already had a memo'd
list inside of it.