Add a fix for centered elements getting put on half pixels.
When you center an element with CSS (either via margin: 0 auto; or
flexbox or otherwise), if its container is not an even number of pixels
wide (or tall, although centering vertically is a bit less common), it
will end up on a half pixel, and can result in things looking blurry as
they get anti-aliased more than usual.

Generally in our design system, we're safe as long as we make the "root"
of our layouts an even number of pixels, as we only use even values for
spacing (or values that combine to be even). As such this commit adds
something that will return `1px` if the dimension is odd, or `0px` if
even, attached at document.body as a CSS custom property. Anything that
needs to know that value (generally, roots that use `margin: 0 auto`)
can do something to adjust their size (padding is often fine, but a
transparent border works too).

This commit also fixes a couple other things that were in the immediate
vicinity while doing this:

- I adjusted the scrim to not appear over the app bar, since we don't
cover that with activity overlays and more and it's nice to be able to
drag it when an overlay or dialog is up. This required some other
adjustments so that didn't make the splash page download dialog look
weird (more CSS properties hooray :) )

- I colored the login page's draggable area so that people know it's
there. Looks a bit tight with the logo but we need to clean that stuff
up anyway.