Add the beginnings of a party system. (#588)
* Add the beginnings of a party system.

This PR adds the server-side party service (and API handlers) for our
party system. Currently, only the following three actions are
implemented:
  - Send an invite to someone
  - Decline an invite (or cancel an invite you sent)
  - Accept an invite

Not sure how many more actions will be added to this PR, before we
decide it's ok to merge it. But I'm guessing at least the "Leave the
party" action as well.

NOTE: Some things will most likely change since we still haven't defined
a clear flowchart for all the possible scenarios, nor did we settle on
a UI/UX yet.

* Set up a FakeNydusServer and write a test that checks that a client has been subscribed.

Includes a fix for a bug with ClientSocketsGroup that resulted in all
clients being subscribed to a route with /clients/undefined/undefined
(still not totally sure what is going on there, something weird with
object prototypes).

* Add a few more party-service tests.

* Update party-service tests according to the suggested comments.

* Put the parties stuff behind a feature flag.

* Disallow inviting yourself to the party.

* Convert the party record to JSON before sending it over the wire. Also, send the leader's name, instead of ID to the user being invited.

* Move the inviting-yourself check to the party-service and add a test for it.

* Send all relevant information about leader to the clients.

* Notify the user when their party invite gets removed, either by them
accepting it or declining it.

* Send the party invites and members as a regular array to the client.

Co-authored-by: Travis Collins <travis@tec27.com>