Add a notifications panel and use it for email verification.
This moves us to the mini button design for settings (and adds
notifications there). Friends list is as-of-yet not implemented so I
didn't add that button for now.

Notifications are fairly generic and expected to have their UI and
behavior implemented client-side, can be generated client-side (provided
they are transient, don't generate ones that need persistence on the
client!). At the moment we just have one type of UI (Actionless) but
more can be added.

These are structured sort of like the chat messages are, as a
discriminated union. We have two layers of that (an interface and a
Record version), which I think is unfortunate, but it is what it is.
Right now the only member of that union is EmailVerification.

Notifications can be unread, in which case their presence adds an
indicator to the button and highlights that notification in the list.
When the list is dismissed, any notifications that were in it are marked
as read (this UX may need some tweaking, but for the moment I think this
is okay). Users can also clear notifications (atm this is only a clear
all function, but we probably want to let them clear individual ones at
some point).

Notifications have an `id` field which should be unique to that
particular notification. For email verification this is always the same
string (we only want to show one of those notifications), but for e.g.
party invites we'd want to generate a new ID for each invite
notification. This mainly ensures that for server-generated
notifications, we A) don't show duplicate notifications on reconnect,
and B) can communicate changes to particular notifications (e.g. "this
user how now read notification XYZ").

This commit also fixes up our usage of react-redux hooks so that they
actually use proper types and ensure we only dispatch valid actions.