Fix title changing on route changes, properly decode chat + whisper names from URL in title code.
Because we don't put route state in the store any more (or subscribe to
it anywhere in MainLayout), MainLayout doesn't get re-rendered on route
changes unless some other state changes. This means things like the
title can be out of date if they don't correspond to a state change
(e.g. moving between a chat channel and whisper). To fix this, I added a
new component specifically for the title that subscribes to location
changes.

Fixes #654.