Convert our form wrapper to TS.
These types and such are pretty complex and ugly and most of that is due
to HoCs being quite complex to type correctly. Things should *generally*
be safe but using this stuff is probably annoying (I can't seem to get
it to work as a @decorator in TSX files, have to split it into a class +
a const assignment, and it's generally pretty cumbersome even then).
Probably we should replace these with some hooks at some point, would
probably be a lot more convenient and definitely easier to type
correctly.