Add a graph of past matchmaking queue size for admins.
This uses visx to render the graph, and requests data since Monday
(UTC). Ideally we'd request more than that, configurably, and overlay
the weeks over each other so we can compare, but I haven't done that
yet (although the API supports it).

We may need to condense the data before display it, as it's quite a lot
of points and seems to be kind of laggy, but it might be fine in
production mode?