Remove online/offline messages from chat and whispers.
After experimenting with not having these, I definitely think these are
from a bygone era of chat, and Discord shows that they're not really
necessary/desirable. The maybe useful info is probably when someone was
last online, but we can show that elsewhere without polluting the
message list.

This also includes an attempted fix for the "unscrollable message list"
problem I've seen occasionally in prod (by triggering a scroll update on
initial mount).