Make map thumbnails clickable in the map browser.
Now opening the map browser directly and clicking on a map thumbnail
will open the details dialog of that map. Opening the map browser
through the lobby-creation flow will still select that map for the
lobby. It would be nice if there was some way to visually distinguish
from these two actions though, maybe overlay some icon over the
thumbnail?