Add a way of declaring HTTP APIs that doesn't involve globbing a whole directory.
This uses tsyringe's injectAll/resolveAll to inject a bunch of things
that are registered with a particular token, and adds a helper decorator
to assign that token consistently. Should allow us to split code on
features again instead of having a folder for all API stuff, a folder
for all DB stuff, etc. It also allows you to have injected dependencies
for your API as need be (although probably a lot of those should be in a
service instead of directly on the API layer).

This doesn't move the existing ones (other than the matchmaking debug
API, which I used to test this change), they can be migrated as
needed/desired.