Refactor the handling of chat message types.
This is the first (out of many I'm sure) steps that refactors the chat
stuff to make it more organized and with less repetitions. This PR
organizes and refactors a bunch of stuff related to message types.

The idea was simple:
  - Everything that is common to all messaging-related services (chat,
    whispers, lobbies, parties) should go in the "messaging" folder
  - Everything else that's specific to a particular message service goes
    in their respective folders

So related to the message types, in the messaging folder there's only
basic text message type defined that's used in all the services, while
chat and lobbies (whispers don't have any yet) specific message types
are defined in their own folder.

Besides definition of message types, this PR applies the same principle
for messages layout as well. Now each messaging-related service is in
charge of rendering its own messages. `MessageList` component has been
extended with an additional property which allows its users to define
how to render their message, with `MessageList` only rendering the
common message types (e.g. text message).

Lobby has been refactored to take advantage of this new setup.