Add table + permissions for managing rally-point servers dynamically.
This also includes some TSification of permissions-related things (and
fixing up some of the auth record stuff that was earlier on in my
"figuring out how to use immutablejs with TS" journey :) ).

The permission doesn't yet allow you to do anything useful, as I need to
write the admin interface + hook up notifications and things for
clients. The plan is that we'll use the new table instead of the
environment variables, notify clients as the table changes, and have
clients ping servers as needed (as they move towards gameplay activities
rather than on startup as it is now). In dev mode, the server will start
up it's local rally-point server if there are no enabled server entries
in the DB.