Implement a hooks version of our form validation stuff.
I also took the opportunity to tighten up the types on the
model/validators here (so we can use standard Partial stuff instead of
having to worry about null/false values as well). This is a bit complex
as far as hooks usage goes, but a lot less complex as far as TypeScript
types.

Also includes a bugfix for something in the HoC version since I noticed
it (we were only ever notfying the first deferred, I think).