Remove batch updates from redux store.
This causes some weird interactions with events outside the store
(especially wouter's hooks) so it seems safer to not do this. I don't
think it was really buying as a lot either, as we don't tend to dispatch
events super frequently and there are already many layers to prevent
unnecessary re-renders.