Double the game load timeout.
The previous (30 second) timeout was actually being hit by a few people
that likely would have loaded successfully. Since its purpose is to
catch cases where someone has failed to load in a way that wasn't really
reported as an error (e.g. their game isn't making progress towards
loading but it also isn't failing completely?), it was a bit too low.
Ideally our game loading code would have a better idea of when progress
was being made so we wouldn't need a general timeout for the whole
thing, but in lieu of writing a new version of the game loader to make
that happen, we'll just bump the timeout up a bit for now.