Use the infinite-scroll-list component in chat channels and whispers.
Infinite-scroll-list component is now extended to support inserting data
both at the beginning and the ending of the list. At the same time even!
For now we're just inserting at one end of the list, but in future, we
might use it to jump into a middle of some data, that can dynamically
load both the previous data and the next data.

Also refactors a bunch of stuff with regards to fetching messages in
chat channels and whispers.