Add a typesafe way to use Electron IPC.
This adds something similar to how TypedEventEmitter works, but split
across 3 interfaces (one for `invoke`d things, one for events the
Renderer can send, and one for events the main process can send).
There's a wrapper for each of the different ways IPC can be used in
Electron (Renderer, main process handling events, main process sending
events) that type properly.

A few things I don't like about this approach that maybe could be
improved:

- It'd be nice to have interfaces with "real" functions, at least for
the invokes, instead of calling `invoke` and passing a string
- All events have to be declared in one place, it'd be nice to be able
to split these into different classes or whatever and have the
registration code run there instead of in `app.ts`

This commit also includes some typings for various socket-handlers
(namely Chat and Lobbies). These typings are incomplete and are mostly
what could be derived from the surrounding code, but may be an okay
example of how to write typed socket handlers going forward. (Notably
they also don't share types with the server code, which would be
desirable if/when we get around to it)

Fixes #603.