Add a message to the chat when a day changes between any two messages. (#669)
Also reorganizes the message layout stuff a bit. Now a
common-message-layout file contains the implementation of the messages
that are common to all chat services, and message-layout file is a mixed
bag of primitives which can be used to style different kinds of
messages. The organization of this file might need some further
organization at some point, but should be fine for now.