Create a new version of Popover to simplify usage.
The previous version of Popover was getting quite complex/hairy because
of a lack of clear ownership of things like positioning, transitions,
etc. There was a lot of bi-directional state flow, which made it very
hard to handle things like safe areas, scrolling content, etc.

The new version simplifies things by expecting the anchor position to be
passed in (instead of calculating this based on an anchor element),
removing support for separately scaled axes and separate transitions for
content vs background (this honestly made these things feel *too*
separate instead of like a single element, and was barely noticeable for
all the effort/code it took), and using CSS custom props for various
things.

The old Popover still exists as `LegacyPopover` as of this commit, and
only the dev Popover page has been moved over to the new one. I'll
update usages/delete LegacyPopover in follow-up commits.

There may be some other things that'd be nice to have (such as spring
customization), but we can deal with those as we need them.