Move Menu over to the new Popover implementation.
Lots of stuff depended on that, directly or indirectly, so a ton of
stuff changed here (and a lot of class components needed to move to
functions to get useAnchorPosition. In theory we could write a class
component shim for this hook but it seemed easier to just move the
components, and *most* were small anyway).

Lots of typescript conversions here as well to ensure the moves were
safe (ish).

This also includes an added feature to Popover to allow for some
customization of the spring transition (so Select could do a purely
vertical one).