Only load skins that are visible based on user's settings
SC:R would load all carbot and standard HD sprites on startup, as you can
change between them in the main menu (This does happen even if the skin
isn't bought though). We obviously don't have such main menu, so loading
carbot when it's not enabled - or non-carbot when carbot is enabled is just
waste of time and memory.

Additionally if the user has opted to not show skins, we can skip loading
the few preorder sprites as well.