Rewrite button classes with hooks, add ripples (finally).
Our Button classes were all subclassing a root one, mainly to manage
state in the same way, which seemed like a prime usecase for hooks. I
moved the state management parts to a single hook (useButtonState) and
kept the new replacement classes (RaisedButton, TextButton, IconButton)
separate otherwise, so it's easier to style them specifically as needed.

This should make it way easier to customize button content/visuals as
needed, and to reuse the state management across a bunch more UIs (I was
able to use it for a rework of the race picker buttons, for example).

The other big piece of this commit is a ripple implementation, which
we've been meaning to build forever :) This mostly uses details from the
MDC Web implementation, reworked to operate well in a React environment.
I skipped implementing touch events for the moment because A) we don't
support them elsewhere and B) it seemed to add a lot of complexity. This
implementation *should* work for keyboard events as well, but the
buttons don't have it wired through yet (I'll do that in a follow-up).

Also as follow-up fodder: a utility function that returns the necessary
props to wire everything through to the RippleController. This would be
nice because it'd make it easy to use ripples on things like cards, list
entries, etc. that might not be using `useButtonState`.