Refactor the profile components so they're easier to extend. (#674)
This PR creates a ConnectedUserProfileOverlay component which allows the
usage of profile overlay without needing to send pre-made actions to it.
Threading required state/callbacks through the chat-channel components
just to create the user's profile overlay actions was a bit of a pain,
and was especially annoying once we started adding more actions to that
menu that depend on other application state, e.g. (the party invite actions
change depending on whether the user is in your party).

This makes it possible to use the user profile overlay at many different
places (something that we might start doing soon, e.g. in chat
messages), without a lot of boiler plate. However, this is done at the
expense of the flexibility in terms of which actions are displayed in
the overlay. If more flexibility is needed in certain scenarios, it is
still possible to use the un-connected user profile overlay component
and provide the actions manually to it.

Another thing that is left as a TODO in this PR is to be able to
configure the position of the user profile overlay through props,
something that will probably be required once we start showing this
overlay at other places.

Finally, this PR moves all the profile components to Typescript, so
that's nice.