Rework tabs to use hooks and add ripples, rework the indicator transition a bit.
This has some a race condition with transitions that scale its ancestors
(namely dialogs) where will sometimes receive a bad initial position. I
think an approach that would avoid this would be to place the indicator
inside the active tab element instead, and transition between positions.
However, that's a larger change than what I did here, and we have newer
mocks for tabs that don't have this indicator anyway. For now I think
the occasional bug is fine, and we'll move to the new visuals shortly.