Move to new blue background styles.
This commit should fix *most* things, I left the ScrollableContent
styles broken because it makes them more obvious when we're trying to
remove them.

This also adjusts various screens that used to be center aligned when
window width was large to be left aligned. May need some other visual
changes with that, but we can at least try it out.