Give popovers a default border-radius, which they should have had before, whoops.
At some point it'd be really nice to be able to customize the content
container here as well (or maybe we should just render the content
directly there? I think I didn't do that before because some things may
have needed a background and I didn't want to have to check everything).