Start storing user info in a reducer.
This will allow us to do local username->id conversions (and vice
versa), as well as store the results of any remote lookups we do.
Ideally this means we can start to replace all the places we use
usernames currently with user IDs instead (which don't change ever, and
thus a much better choice!).

A bunch of things still communicate only names (chat, whispers, lobbies,
for starters), and need to be updated to communicate more. We also need
a general purpose API for doing username -> ID conversions and looking
up basic info about a user. The info stored in this reducer atm is very
incomplete, I think it'd also be nice to have current status (online,
offline, ingame, etc.), avatar, title, and probably to be able to load
this stuff kind of piecemeal. Implementation of that TBD.