Try out immer for the user reducer.
ImmutableJS has proven to be fairly convoluted in a lot of TypeScript
usages, along with having a pretty weird API relative to most JS stuff.
The `any` typings of `setIn` and the like were particularly aggravating.

Immer makes it so that we could use the same TS interfaces for our state
that we use outside of reducers, which is appealing, and most of the
usefulness of Records goes away when we have `interface` anyway.

Since it's only a 3kb dependency, seemed worthwhile to pull in and try
out for something. I don't think the resulting reducer code here is
necessarily better or worse by more than a bit, but the simpler types
are definitely nice.