Move the whispers stuff away from using usernames internally.
Whispers are kind of in a weird place because we still want to use the
usernames in all the routing-related paths, which means it's still
preferable to use the usernames in various data structures (e.g. in the
reducer).

To be honest, we could probably get away with using usernames everywhere
(which makes this PR kinda useless huh), since we only need user IDs in
a couple of places, like in each whisper message.