Load user data if it's not already present when viewing a profile.
This includes a few changes to try to introduce better patterns:

- Adds an HTTP API Endpoint wrapper that allows us to type the return of
an API function, as well as route/body inputs. The former has annoyed me
a lot (we often have to create an intermediary object just to give it an
enforced type), so hopefully that cleans things up a bit.

- Allows dispatched Promises to return null/undefined to signal that
nothing should be dispatched. This was something I thought I needed for
this commit but didn't end up needing, but it still seems like a useful
change.

- Attempts a new pattern for localizing error handling while still
pushing/retrieving data through Redux. This is currently localized to
the profile action-creators file, but we can move it somewhere common if
it works out. The idea here is that often errors are really only
relevant to the UI component that dispatched the original request, while
the data retrieved is relevant everywhere. Before, we often stored these
errors globally, and thus we had to do things like generate request IDs
and match these IDs up to the error stored in the store, and it was...
not great.

This setup calls a callback with the error, allowing the UI component to
store it in local state, and ensures 100% that the relevant component
sees it. It also takes in (optionally) an AbortSignal so that requests
can be aborted if they're no longer relevant (e.g. if the user navigates
away).