Handle user changing a party.
This can happen when a user gets invited to a party while they were
already in a different party. By accepting the invite to a new party,
they will immediately leave the old party and be put in a new party.

TODO: The transition from one party to another might be a bit jarring,
so maybe eventually we want to display some kind of a confirmation
dialog before doing it?