Detect links in text messages and make them clickable.
This does some basic link detection in text messages, similar to what
Discord does. One thing that should probably be added as soon as
possible as well is some kind of a warning message that we show to users
about opening untrusted links.

But then we should also probably remember their choices for each
specific domain, and that's a whole thing now ^^.

Additionally, the component that is implemented here to parse the text
messages could be extended to parse other types of content (e.g.
mentions, links to lobbies, maps, etc.). Regex for each of those should
be written separately, which should be easier since we can create custom
schemes for each one, and then combined into a single regex so the text
is parsed only once. I don't know yet if this is possible or how it will
work, but we'll see when we get there :).