Display the actual reason why accepting a party invite failed.
These are more or less the only errors that are likely to happen to
users in their normal usage of the party system. Now we display the
actual reason to them if they try to accept an invite to a party that
doesn't exist anymore, or if the party is currently full.