Refactor the matchmaking preferences to support multiple matchmaking types.
This PR also refactors the terminology of "preferred maps" to a more
general "map selections". This means that "map selections" may be used
for any kind of map system we might have (e.g. in 1v1/2v2 matchmaking it
might mean "vetoed maps", while in 3v3 matchmaking it might mean "maps
selected to play on".

We're also adding 2v2 matchmaking type in here to make things easier
with typescript. Actual implementation of the 2v2 matchmaking will come
in subsequent PRs.

Finally, we've stopped rendering the tabs in find-match overlay for
types we don't support yet, e.g. 3v3. Keeping the tab showing while not
having the support for it is more trouble than it's worth, and besides,
hopefully we'll add 3v3 matchmaking soon after the 2v2 one ^^.