Don't close the game when failing to send window move message to app
The async task handling messages from game thread closed due when it was
not able to handle the message, and, perhaps unexpectedly, caused the
main game state task to close as well due to all tasks sending messages
had closed.
Now the game thread forwarding task will not exit on failure to forward
a message, which is likely how it should've worked in the first place.