Detect links in text messages and make them clickable. (#727)
This does some basic link detection in text messages, similar to what
Discord does. One thing that should probably be added as soon as
possible as well is some kind of a warning message that we show to users
about opening untrusted links.

But then we should also probably remember their choices for each
specific domain, and that's a whole thing now ^^.

Additionally, the component that is implemented here to parse the text
messages could be extended to parse other types of content (e.g.
mentions, links to lobbies, maps, etc.). Regexes for each of those should
be written separately, which should be easier than trying to combine all
of the functionality into a single expression.

There are known bugs with the URL regex relating to IPv6 addresses, an
issue will be filed to fix these (they don't seem like a super-common
usecase so having them broken for a bit doesn't seem like the hugest
deal).

Co-authored-by: Travis Collins <travis@tec27.com>