Fix issue with ripples queueing deactivation animations as they unmount.
We now cancel any requestAnimationFrame calls on unmount (we were doing
this for timeouts previously, but I sort of assumed it wouldn't be
possible to get a timing such that requestAnimationFrame did stuff, and
if it was, the setState wouldn't be particularly harmful. I guess we
found a way and React disagrees about the harm, so, fixed!).

Fixes #747.