Move to using Reflect metadata for Http APIs instead of storing data in a class/using constructor calls.
This is what I originally wanted but I didn't really want to figure it
out at the time, makes the boilerplate a bit less heavy. Ideally I'd
like to get rid of the `applyRoutes` function as well (and the interface
as a whole) but I need to write decorators for the methods first.