Add method decorators for specifying an API route and middleware.
These serve much the same purpose as `apiEndpoint`, minus validations. I
think validations could be included but A) I'm not sure it'll be super
readable/formatted well, and B) it may be hard to make the typings for
the method work (although I think I already did most of the hard stuff
here so maybe not).

This commit will be a bit spammy because most of the HttpApi classes
still use the `applyRoutes` method instead of decorators, and not having
any decorators in these classes causes a warning. I'll move the rest of
the stuff over ASAP to silence those.

One other thing we need is a `withMiddleware` sort of thing that works
as a class decorator. I'll probably add that next, prior to migrating
anything, to minimize the migration work.