Add a tsyringe child container for each request. [NOT READY]
This could be used to scope the lifetime of various objects to the
lifetime of a request, while allowing them to be lazily created, cached
upon creation, etc.

Originally I wanted to use this container to inject the method
parameters of @httApi methods, but this turns out to not be doable with
tsyringe (it only stores metadata keyed by class, rather than
class+method). I think this approach could still be useful for
constructing objects inside those methods, but it's definitely not worth
the code necessary just for e.g. returning the current user ID.

Because of that, I'm pushing this up for reference later, once we have a
more concrete use case/need for this type of DI.