Move to a map veto system rather than preferred maps.
The previous system was rather complex to explain and understand,
leading to a lot of spurious bug reports and confusion. It also made it
hard to construct a good map pool: since all maps had a shot at being
chosen, we could include "weird" maps easily without making it very hard
for new players.

The goal of the previous system was encouraging map diversity on the
ladder versus systems without matchmaking like ICCup, which tend to
devolve into 1 or 2 maps only. Since then, SC:R has shown that a veto
system is mostly fine for map diversity relative to that, and this
system will allow us to construct more interesting map pools.

This commit changes the Find Match UI to show the entire pool upfront,
instead of on a separate page, and allows users to manipulate the maps
directly. It also cleans up the Find Match code a lot (still more to be
done there) to simplify handling multiple matchmaking types.

There are some eager initializations of the map pool data in here to
facilitate showing this form UI, but I think ideally we'd be able to
request that on the fly as well, so I'll probably build that next.