Add redux middleware to automatically tag actions with the current monotonic time.
With the data normalization approach to our reducers, it seemed nearly
impossible to ensure we'd always have the current time when we needed to
store things (e.g. maps, when maps can come from any of 10 different
requests). Rather than repeat the same logic and update all these action
creators to do the right thing, we just provide this data on every
action now.

window.performance.now() should be pretty inexpensive to call so I don't
foresee this causing a lot of issues. There are a few actions that
utilize that value already and can probably be adjusted to use the
system info instead, but I haven't attempted to fix them for this
commit.