Use population estimates to pick better matchmaking search ranges.
Now when players queue, we use our collected population estimates to see
if we think they're likely to (eventually) find a match (when they reach
their max search bounds). If we see that they aren't likely to, we
increase the bounds until our estimates say it's likely.

Bound increases are done at the same speed as before (so expanding a
search to [0, 3000] will still take ~20 minutes). I think this is mostly
fine, as that match isn't really a desirable one, and we can thus say at
20 minutes we'll give it as a last resort. (I think in the long run we
don't really want to give that match *at all* but we gotta build a
playerbase first :) )

If a player reaches their max range and still hasn't found anyone, we'll
re-check to see if that range is sensible given our estimates (at the
same rate that estimates are recalculated). This means that ranges tend
towards [0, 3000] over time, if the matchmaker doesn't have a constant
influx of players. Any 2 players in the queue will always find each
other eventually, if nobody else queues.