Decrease the amount of time that we prioritize an ideal match to 3 iterations (18 seconds).
Before we tried this for 60 seconds. In a perfect world, we'd have the
player base to support that, but at the moment we're dealing with a
small player base and competing with a matchmaking system that will
deliver tons of bad matches within 30 seconds, so not even increasing
our search bounds until 60 seconds seems a bit too long. We'll still
*hope* for a great match in those first 18 seconds, but after that we'll
at least start bumping up the search range immediately.