Increase search bounds twice as fast (100 seconds -> 50 seconds max).
Similar to the last commit, this change attempts to speed up matching in
a low pop environment. We prioritize finding games a bit more over
finding a perfect game. In the long run we may be able to tweak this
back to the initial value a bit more, but we probably need more data to
see that :)

This will mean anyone within your (normal, non-pop-adjusted) range is
findable within 78 seconds.