Fix injector to work for process reuse (I think?).
I'm not sure that we do anything in here that is actually dangerous for
process reuse (the class initialization looks to be the main
possibility, to me), but I'm not sure this process reuse really matters
for us anyway. We don't utilize this module from the renderer process
(only the main process, which won't be reused), and we also don't create
multiple renderer processes anyway atm. So, overall, I think the risk is
quite low, but since Electron 14+ requires this feature to be on, I made
a small attempt to correct things.

(VSCode also reformatted everything, which is where the bulk of the
delta comes from)